i am giving up on CSTs 😔
rust-analyzer's generated AST code doesn't work for our purposes, because it assumes that there will never be a nested enum, only a struct -> enum -> struct nesting.
i don't grok ungrammar well enough to be able to generate the code i want; and this is kind of a bonus feature anyway, in practice "not reordering statements" will be enough for 99+% of our needs.

this records my progress just in case i want to look at it later